[TOPIC-BLE-LLCP]: Bluetooth: controller: llcp: split out control procs
The control procedures are splitted out in their own file, making it
easier to merge in the new LLCP
